corecting bug

+++ drive.py
@@ -37,7 +37,7 @@ def telemetry(sid, data):
     imgString = data["image"]
     image = Image.open(BytesIO(base64.b64decode(imgString)))
     image= image.crop((0,60,320,135))
-    image_array = (np.asarray(image)
+    image_array = (np.asarray(image)).astype(np.float32)
     transformed_image_array = image_array[None, :, :, :]
     # This model currently assumes that the features of the model are just the images. Feel free to change this.
     steering_angle = float(model.predict(transformed_image_array, batch_size=1))
